+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Ability to add interesting cultivars to a species.
 - Ability to add maintainers' notes to a species.
 - Icon to see public pages for a species in its edit page.
+- Summernote added for maintainer's notes.
 
 # 0.4.0 - Pre-release (2019-08-25)
 - Fix comestible usage information on show page.